requirements.txt
weaviate-client>=4.0.0
gradio>=4.0.0
langchain>=0.0.267
pypdf>=3.15.1
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
openai>=1.0.0
